Add functionality: It is now able to provide more membership details, able to join the community and view other community members

+++ index.php
@@ -1,5 +1,25 @@
 <?php
     session_start();
+    
+    $fetch = null;
+    if(isset($_SESSION['id'])){
+        try{
+        require_once('connect.php');
+        $query = "SELECT * FROM user_info_more WHERE id = :id;";       
+        $statement = $db->prepare($query);
+        $statement->bindValue(':id', $_SESSION['id']);
+        $statement->execute();
+        $fetch = $statement->fetch();
+        if(isset($fetch['id'])){
+            $_SESSION['isJoin'] = 1;
+        }else{
+            $_SESSION['isJoin'] = null;
+        }}catch(PDOException $e){
+            echo $e->getMessage();
+        }
+        
+    }
+        
     if(isset($_SESSION['isLogin'])){
 ?>
 <html lang="en">
@@ -18,13 +38,26 @@
     </header>
     <main>
     <?php         
-        echo '<h2>Welcome! '.$_SESSION['user_first_name'].' </h2>';          
+        echo '<h2>Welcome! '.$_SESSION['user_first_name'].' </h2>';
+        if(isset($_SESSION['isJoin'])){
         echo  '<nav>'.
-            '<ul>'.
-                '<li><a href="view.php?id="'.$_SESSION['id'].'">Modify Your Information</a></li>'.
-                '<li><a href="session_destroy.php">Logout</a></li>'.
-            '</ul>'.
-        '</nav>';
+                '<ul>'.
+                    '<li><a href="view.php?id="'.$_SESSION['id'].'">Modify Account Information</a></li>'.
+                    '<li><a href="join.php">Modify Personal Information</a></li>'.
+                    '<li><a href="overview.php">View Network</a></li>'.
+                    '<li><a href="session_destroy.php">Logout</a></li>'.
+                '</ul>'.
+               '</nav>';
+        
+        }else{
+            echo'<nav>'.
+                '<ul>'.
+                    '<li><a href="view.php?id="'.$_SESSION['id'].'">Modify Account Information</a></li>'.
+                    '<li>Let\'s connected and <a href="join.php">Join Community</a></li>'.
+                    '<li><a href="session_destroy.php">Logout</a></li>'.
+                    '</ul>'.
+                '</nav>';
+        }
     }else{    
         echo '<h2>Welcome to the community!</h2>'.
              '<h3>Let\'s Join and Incorporate!</h3>';
@@ -39,10 +72,10 @@
 
     }?>
 
-        
+    <div><a href="review.html">Final Project Review</a></div>
     </main>
     <footer>
-      <p> &copy; 2020 COMP1006 - Lab Nine </p>
+      <p> &copy; 2020 COMP1006 - Final Project </p>
     </footer>
    </div><!--end container-->
   </body>
